refactor: PathFinder interface 적용
