+++ src/test/java/nextstep/subway/path/domain/PathFinderTest.java
@@ -46,10 +46,10 @@ public class PathFinderTest {
     @DisplayName("최단 경로 탐색")
     void 최단경로탐색() {
         //given
-        PathFinder pathFinder = new PathFinder(Arrays.asList(신분당선, 이호선, 삼호선));
+        DijkstraPathFinder dijkstraPathFinder = new DijkstraPathFinder(Arrays.asList(신분당선, 이호선, 삼호선));
 
         //when
-        Path path = pathFinder.findShortestPath(교대역, 양재역);
+        Path path = dijkstraPathFinder.findShortestPath(교대역, 양재역);
 
         //then
         assertAll(
@@ -62,10 +62,10 @@ public class PathFinderTest {
     @DisplayName("최단 경로 탐색 실패 : 출발역과 도착역이 같은 경우")
     void 최단경로탐색_실패_출발역과_도착역이_같음() {
         //given
-        PathFinder pathFinder = new PathFinder(Arrays.asList(신분당선, 이호선, 삼호선));
+        DijkstraPathFinder dijkstraPathFinder = new DijkstraPathFinder(Arrays.asList(신분당선, 이호선, 삼호선));
 
         //then
-        assertThrows(IllegalArgumentException.class, () -> pathFinder.findShortestPath(교대역, 교대역));
+        assertThrows(IllegalArgumentException.class, () -> dijkstraPathFinder.findShortestPath(교대역, 교대역));
     }
 
     @Test
@@ -75,10 +75,10 @@ public class PathFinderTest {
         Station 부평역 = createStation("부평역");
         Station 인천시청역 = createStation("인천시청역");
         Line 인천선 = createLine("신분당선", "rbg-red-600", 부평역, 인천시청역, 10);
-        PathFinder pathFinder = new PathFinder(Arrays.asList(신분당선, 이호선, 삼호선, 인천선));
+        DijkstraPathFinder dijkstraPathFinder = new DijkstraPathFinder(Arrays.asList(신분당선, 이호선, 삼호선, 인천선));
 
         //then
-        assertThrows(IllegalArgumentException.class, () -> pathFinder.findShortestPath(교대역, 부평역));
+        assertThrows(IllegalArgumentException.class, () -> dijkstraPathFinder.findShortestPath(교대역, 부평역));
     }
 
     @Test
@@ -86,9 +86,9 @@ public class PathFinderTest {
     void 최단경로탐색_실패_존재하지_않는_역() {
         //given
         Station 존재하지않는역 = createStation("존재하지않는역");
-        PathFinder pathFinder = new PathFinder(Arrays.asList(신분당선, 이호선, 삼호선));
+        DijkstraPathFinder dijkstraPathFinder = new DijkstraPathFinder(Arrays.asList(신분당선, 이호선, 삼호선));
 
         //then
-        assertThrows(IllegalArgumentException.class, () -> pathFinder.findShortestPath(교대역, 존재하지않는역));
+        assertThrows(IllegalArgumentException.class, () -> dijkstraPathFinder.findShortestPath(교대역, 존재하지않는역));
     }
 }
